osdc,librbd: use int64_t type for qos api

+++ src/librbd/internal.cc
@@ -2482,7 +2482,7 @@ bool compare_by_name(const child_info_t& c1, const child_info_t& c2)
   }
 
   int qos_param_check(ImageCtx *ictx,
-                      int reservation, int weight, int limit, int bandwidth) {
+                      int64_t reservation, int64_t weight, int64_t limit, int64_t bandwidth) {
     if (reservation < -1 && weight < -1 && limit < -1
         && bandwidth < -1) {
       ldout(ictx->cct, 0) << "invalid qos, must be greater than -1." << dendl;
@@ -2506,7 +2506,7 @@ bool compare_by_name(const child_info_t& c1, const child_info_t& c2)
   }
 
   int qos_spec_set(ImageCtx *ictx,
-                       int rsv, int wgt, int lmt, int bdw)
+                       int64_t rsv, int64_t wgt, int64_t lmt, int64_t bdw)
   {
     int r = qos_param_check(ictx, rsv, wgt, lmt, bdw);
     if (r < 0) {
@@ -2525,7 +2525,7 @@ bool compare_by_name(const child_info_t& c1, const child_info_t& c2)
   }
 
   int qos_spec_get(ImageCtx *ictx,
-                       int *rsv, int *wgt, int *lmt, int *bdw,
+                       int64_t *rsv, int64_t *wgt, int64_t *lmt, int64_t *bdw,
                        int *mflag)
   {
     string srsv, swgt, slmt, sbdw;
@@ -2541,7 +2541,7 @@ bool compare_by_name(const child_info_t& c1, const child_info_t& c2)
         *wgt = ictx->cct->_conf->rbd_client_qos_weight;
       }
     } else {
-      *wgt = std::stoi(swgt, nullptr);
+      *wgt = std::stol(swgt, nullptr);
       if (mflag) {
         *mflag |= QOS_FLAG_WGT;
       }
@@ -2553,7 +2553,7 @@ bool compare_by_name(const child_info_t& c1, const child_info_t& c2)
           *rsv = ictx->cct->_conf->rbd_client_qos_reservation;
         }
       } else {
-        *rsv = std::stoi(srsv, nullptr);
+        *rsv = std::stol(srsv, nullptr);
         if (mflag) {
           *mflag |= QOS_FLAG_RSV;
         }
@@ -2570,7 +2570,7 @@ bool compare_by_name(const child_info_t& c1, const child_info_t& c2)
           *lmt = ictx->cct->_conf->rbd_client_qos_limit;
         }
       } else {
-        *lmt = std::stoi(slmt, nullptr);
+        *lmt = std::stol(slmt, nullptr);
         if (mflag) {
           *mflag |= QOS_FLAG_LMT;
         }
@@ -2587,7 +2587,7 @@ bool compare_by_name(const child_info_t& c1, const child_info_t& c2)
           *bdw = ictx->cct->_conf->rbd_client_qos_bandwidth;
         }
       } else {
-        *bdw = std::stoi(sbdw, nullptr);
+        *bdw = std::stol(sbdw, nullptr);
         if (mflag) {
           *mflag |= QOS_FLAG_BDW;
         }
